dev-requirements.txt
apache-airflow[crypto,celery,postgres,hive,jdbc,sendgrid,mysql,ssh]==2.0.1